feat: implement advanced  setting
