+++ src/pages/Typing/components/Setting/index.tsx
@@ -1,74 +1,14 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { Dialog, Switch, Tab, Transition } from '@headlessui/react'
+import { Dialog, Tab, Transition } from '@headlessui/react'
 import classNames from 'classnames'
-import { Fragment, useCallback, useState } from 'react'
-import * as Slider from '@radix-ui/react-slider'
-import { useAtom } from 'jotai'
-import { hintSoundsConfigAtom, keySoundsConfigAtom, pronunciationConfigAtom } from '@/store'
-import styles from './index.module.css'
+import { Fragment, useContext, useState } from 'react'
+import SoundSetting from './SoundSetting'
+import AdvancedSetting from './AdvancedSetting'
+import { TypingContext, TypingStateActionType } from '../../store'
 
 export default function Setting() {
-  const [isOpen, setIsOpen] = useState(false)
-  const [pronunciationConfig, setPronunciationConfig] = useAtom(pronunciationConfigAtom)
-  const [keySoundsConfig, setKeySoundsConfig] = useAtom(keySoundsConfigAtom)
-  const [hintSoundsConfig, setHintSoundsConfig] = useAtom(hintSoundsConfigAtom)
-
-  const onTogglePronunciation = useCallback(
-    (checked: boolean) => {
-      setPronunciationConfig((prev) => ({
-        ...prev,
-        isOpen: checked,
-      }))
-    },
-    [setPronunciationConfig],
-  )
-  const onChangePronunciationVolume = useCallback(
-    (value: [number]) => {
-      setPronunciationConfig((prev) => ({
-        ...prev,
-        volume: value[0] / 100,
-      }))
-    },
-    [setPronunciationConfig],
-  )
-
-  const onToggleKeySounds = useCallback(
-    (checked: boolean) => {
-      setKeySoundsConfig((prev) => ({
-        ...prev,
-        isOpen: checked,
-      }))
-    },
-    [setKeySoundsConfig],
-  )
-  const onChangeKeySoundsVolume = useCallback(
-    (value: [number]) => {
-      setKeySoundsConfig((prev) => ({
-        ...prev,
-        volume: value[0] / 100,
-      }))
-    },
-    [setKeySoundsConfig],
-  )
-
-  const onToggleHintSounds = useCallback(
-    (checked: boolean) => {
-      setHintSoundsConfig((prev) => ({
-        ...prev,
-        isOpen: checked,
-      }))
-    },
-    [setHintSoundsConfig],
-  )
-  const onChangeHintSoundsVolume = useCallback(
-    (value: [number]) => {
-      setHintSoundsConfig((prev) => ({
-        ...prev,
-        volume: value[0] / 100,
-      }))
-    },
-    [setHintSoundsConfig],
-  )
+  const [isOpen, setIsOpen] = useState(true)
+  const { dispatch } = useContext(TypingContext) ?? {}
 
   function closeModal() {
     setIsOpen(false)
@@ -76,6 +16,9 @@ export default function Setting() {
 
   function openModal() {
     setIsOpen(true)
+    if (dispatch) {
+      dispatch({ type: TypingStateActionType.SET_IS_TYPING, payload: false })
+    }
   }
 
   return (
@@ -155,106 +98,10 @@ export default function Setting() {
 
                       <Tab.Panels className="h-full w-full flex-1">
                         <Tab.Panel className="flex h-full w-full  focus:outline-none">
-                          <div className="flex w-full flex-col items-start justify-start gap-10 overflow-y-auto pb-40 pl-6 pr-9 pt-8">
-                            <div className={styles.section}>
-                              <span className={styles.sectionLabel}>单词发音</span>
-                              <div className={styles.switchBlock}>
-                                <Switch checked={pronunciationConfig.isOpen} onChange={onTogglePronunciation} className="switch-root">
-                                  <span aria-hidden="true" className="switch-thumb" />
-                                </Switch>
-                                <span className="text-right text-xs font-normal leading-tight text-gray-600">{`发音已${
-                                  pronunciationConfig.isOpen ? '开启' : '关闭'
-                                }`}</span>
-                              </div>
-                              <div className={styles.block}>
-                                <span className={styles.blockLabel}>音量</span>
-                                <div className="flex h-5 w-full items-center justify-between">
-                                  <Slider.Root
-                                    defaultValue={[pronunciationConfig.volume * 100]}
-                                    max={100}
-                                    step={10}
-                                    className="slider"
-                                    onValueChange={onChangePronunciationVolume}
-                                  >
-                                    <Slider.Track>
-                                      <Slider.Range />
-                                    </Slider.Track>
-                                    <Slider.Thumb />
-                                  </Slider.Root>
-                                  <span className="ml-4 w-10 text-xs font-normal text-gray-800">{`${Math.floor(
-                                    pronunciationConfig.volume * 100,
-                                  )}%`}</span>
-                                </div>
-                              </div>
-                            </div>
-
-                            <div className={styles.section}>
-                              <span className={styles.sectionLabel}>按键音</span>
-                              <div className={styles.switchBlock}>
-                                <Switch checked={keySoundsConfig.isOpen} onChange={onToggleKeySounds} className="switch-root">
-                                  <span aria-hidden="true" className="switch-thumb" />
-                                </Switch>
-                                <span className="text-right text-xs font-normal leading-tight text-gray-600">{`发音已${
-                                  keySoundsConfig.isOpen ? '开启' : '关闭'
-                                }`}</span>
-                              </div>
-                              <div className={styles.block}>
-                                <span className="font-medium text-gray-800">音量</span>
-                                <div className="flex h-5 w-full items-center justify-between">
-                                  <Slider.Root
-                                    defaultValue={[keySoundsConfig.volume * 100]}
-                                    max={100}
-                                    step={10}
-                                    className="slider"
-                                    onValueChange={onChangeKeySoundsVolume}
-                                  >
-                                    <Slider.Track>
-                                      <Slider.Range />
-                                    </Slider.Track>
-                                    <Slider.Thumb />
-                                  </Slider.Root>
-                                  <span className="ml-4 w-10 text-xs font-normal text-gray-800">{`${Math.floor(
-                                    keySoundsConfig.volume * 100,
-                                  )}%`}</span>
-                                </div>
-                              </div>
-                            </div>
-
-                            <div className={styles.section}>
-                              <span className={styles.sectionLabel}>效果音</span>
-                              <div className={styles.switchBlock}>
-                                <Switch checked={hintSoundsConfig.isOpen} onChange={onToggleHintSounds} className="switch-root">
-                                  <span aria-hidden="true" className="switch-thumb" />
-                                </Switch>
-                                <span className="text-right text-xs font-normal leading-tight text-gray-600">{`发音已${
-                                  hintSoundsConfig.isOpen ? '开启' : '关闭'
-                                }`}</span>
-                              </div>
-                              <div className={styles.block}>
-                                <span className="font-medium text-gray-800">音量</span>
-                                <div className="flex h-5 w-full items-center justify-between">
-                                  <Slider.Root
-                                    defaultValue={[hintSoundsConfig.volume * 100]}
-                                    max={100}
-                                    step={10}
-                                    className="slider"
-                                    onValueChange={onChangeHintSoundsVolume}
-                                  >
-                                    <Slider.Track>
-                                      <Slider.Range />
-                                    </Slider.Track>
-                                    <Slider.Thumb />
-                                  </Slider.Root>
-                                  <span className="ml-4 w-10 text-xs font-normal text-gray-800">{`${Math.floor(
-                                    hintSoundsConfig.volume * 100,
-                                  )}%`}</span>
-                                </div>
-                              </div>
-                            </div>
-                          </div>
+                          <SoundSetting />
                         </Tab.Panel>
                         <Tab.Panel className="h-full w-full px-6 py-3 focus:outline-none">
-                          <h1>账号设置2</h1>
+                          <AdvancedSetting />
                         </Tab.Panel>
                       </Tab.Panels>
                     </div>
