+++ src/pages/Typing/components/Switcher/index.tsx
@@ -3,21 +3,17 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { useHotkeys } from 'react-hotkeys-hook'
 import Tooltip from '@/components/Tooltip'
 import { useAtom } from 'jotai'
-import { isLoopSingleWordAtom, isOpenDarkModeAtom, randomConfigAtom } from '@/store'
+import { isLoopSingleWordAtom, isOpenDarkModeAtom } from '@/store'
 import { TypingContext, TypingStateActionType } from '../../store'
 import SoundSwitcher from '../SoundSwitcher'
 import Setting from '../Setting'
 
 export default function Switcher() {
-  const [randomConfig, setRandomConfig] = useAtom(randomConfigAtom)
   const [isOpenDarkMode, setIsOpenDarkMode] = useAtom(isOpenDarkModeAtom)
   const [isLoopSingleWord, setIsLoopSingleWord] = useAtom(isLoopSingleWordAtom)
 
   const { state, dispatch } = useContext(TypingContext) ?? {}
 
-  const changeRandomState = () => {
-    setRandomConfig((old) => ({ ...old, isOpen: !old.isOpen }))
-  }
   const changeDarkModeState = () => {
     setIsOpenDarkMode((old) => !old)
   }
@@ -48,14 +44,6 @@ export default function Switcher() {
   )
 
   useHotkeys(
-    'ctrl+u',
-    () => {
-      changeRandomState()
-    },
-    { enableOnFormTags: true, preventDefault: true },
-    [],
-  )
-  useHotkeys(
     'ctrl+d',
     () => {
       changeDarkModeState()
@@ -85,17 +73,7 @@ export default function Switcher() {
       <Tooltip content="音效设置">
         <SoundSwitcher />
       </Tooltip>
-      <Tooltip content="开关单词乱序（Ctrl + U）">
-        <button
-          className={`${randomConfig.isOpen ? 'text-indigo-400' : 'text-gray-400'} text-lg focus:outline-none`}
-          onClick={(e) => {
-            changeRandomState()
-            e.currentTarget.blur()
-          }}
-        >
-          <FontAwesomeIcon icon="random" fixedWidth />
-        </button>
-      </Tooltip>
+
       <Tooltip content="开关单个单词循环（Ctrl + L）">
         <button
           className={`${isLoopSingleWord ? 'text-indigo-400' : 'text-gray-400'} text-lg focus:outline-none`}
